requirements.txt
astroid==2.3.1
autopep8==1.4.4
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.8
isort==4.3.21
jieba3k==0.35.1
joblib==0.14.0
lazy-object-proxy==1.4.2
lxml==4.4.1
mccabe==0.6.1
newspaper3k==0.2.8
nltk==3.4.5
numpy==1.17.2
Pillow==6.2.0
pycodestyle==2.5.0
pylint==2.4.2
python-dateutil==2.8.0
PyYAML==5.1.2
requests==2.22.0
requests-file==1.4.3
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
soupsieve==1.9.4
tinysegmenter==0.3
tldextract==2.2.1
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
